Move standard implementation to spi package
