+++ airframe-msgpack/shared/src/main/scala/wvlet/airframe/msgpack/spi/OffsetPacker.scala
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.msgpack.io
+package wvlet.airframe.msgpack.spi
 
 import java.math.BigInteger
 import java.nio.charset.StandardCharsets
